requirements.txt
fastapi
uvicorn[standard]
pydantic
sse-starlette
edge-tts
groq
moviepy>=1.0.3
numpy
Pillow
imageio-ffmpeg
python-multipart
